refactor: 💡 migrate credential-library off fragments
