--- addons/api/app/models/fragment-credential-library-attributes.js
@@ -1,1 +1,0 @@
-export { default } from 'api/models/fragment-credential-library-attributes';
