--- addons/api/app/serializers/fragment-credential-library-attributes.js
@@ -1,1 +1,0 @@
-export { default } from 'api/serializers/fragment-credential-library-attributes';
