--- addons/api/tests/unit/models/fragment-credential-library-attributes-test.js
@@ -1,19 +1,0 @@
-import { module, test } from 'qunit';
-import { setupTest } from 'ember-qunit';
-
-module(
-  'Unit | Model | fragment credential library attributes',
-  function (hooks) {
-    setupTest(hooks);
-
-    // Replace this with your real tests.
-    test('it exists', function (assert) {
-      let store = this.owner.lookup('service:store');
-      let model = store.createRecord(
-        'fragment-credential-library-attributes',
-        {}
-      );
-      assert.ok(model);
-    });
-  }
-);
