+++ ui/admin/app/components/form/credential-library/index.hbs
@@ -56,17 +56,17 @@
   <form.input
     @name='vault_path'
     @type='text'
-    @value={{@model.attributes.path}}
+    @value={{@model.path}}
     @label={{t 'resources.credential-library.form.vault_path.label'}}
-    @error={{@model.errors.attributes_path}}
+    @error={{@model.errors.path}}
     @helperText={{t 'resources.credential-library.form.vault_path.help'}}
     disabled={{@model.isSaving}}
     readonly={{false}}
     as |field|
   >
-    {{#if @model.errors.attributes_path}}
+    {{#if @model.errors.path}}
       <field.errors as |errors|>
-        {{#each @model.errors.attributes_path as |error|}}
+        {{#each @model.errors.path as |error|}}
           <errors.message>{{error.message}}</errors.message>
         {{/each}}
       </field.errors>
@@ -81,8 +81,8 @@
     <:body>
       <form.radioGroup
         @name='http_method'
-        @selectedValue={{@model.attributes.http_method}}
-        @changed={{fn (mut @model.attributes.http_method)}}
+        @selectedValue={{@model.http_method}}
+        @changed={{fn (mut @model.http_method)}}
         @inline={{true}}
         disabled={{@model.isSaving}}
         as |radioGroup|
@@ -98,9 +98,9 @@
     <form.textarea
       @name='http_request_body'
       @type='text'
-      @value={{@model.attributes.http_request_body}}
+      @value={{@model.http_request_body}}
       @label={{t 'resources.credential-library.form.http_request_body.label'}}
-      @error={{@model.errors.attributes_http_request_body}}
+      @error={{@model.errors.http_request_body}}
       @helperText={{t
         'resources.credential-library.form.http_request_body.help'
       }}
@@ -108,9 +108,9 @@
       readonly={{false}}
       as |field|
     >
-      {{#if @model.errors.attributes_http_request_body}}
+      {{#if @model.errors.http_request_body}}
         <field.errors as |errors|>
-          {{#each @model.errors.attributes_http_request_body as |error|}}
+          {{#each @model.errors.http_request_body as |error|}}
             <errors.message>{{error.message}}</errors.message>
           {{/each}}
         </field.errors>
